Rename database service
